+++ semanticz-connector-fuseki/src/main/java/zone/cogni/semanticz/connectors/fuseki/FusekiSparqlService.java
@@ -1,6 +1,24 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 package zone.cogni.semanticz.connectors.fuseki;
 
-import org.apache.http.HttpHeaders;
 import org.apache.jena.query.QueryExecution;
 import org.apache.jena.query.QueryExecutionBuilder;
 import org.apache.jena.query.ResultSet;
@@ -23,6 +41,7 @@ import java.net.http.HttpRequest.BodyPublishers;
 import java.nio.charset.StandardCharsets;
 import java.util.function.Function;
 
+import static zone.cogni.semanticz.connectors.utils.Constants.CONTENT_TYPE;
 import static zone.cogni.semanticz.connectors.utils.HttpClientUtils.execute;
 
 /**
@@ -56,7 +75,7 @@ public class FusekiSparqlService implements SparqlService {
       request = HttpRequest
           .newBuilder(URI.create(sparqlUrl))
           .POST(BodyPublishers.ofFile(file.toPath()))
-          .header(HttpHeaders.CONTENT_TYPE, config.getTurtleMimeType() + ";charset=utf-8")
+          .header(CONTENT_TYPE, config.getTurtleMimeType() + ";charset=utf-8")
           .build();
     } catch (FileNotFoundException e) {
       throw new RuntimeException(e);
@@ -77,16 +96,10 @@ public class FusekiSparqlService implements SparqlService {
         .newBuilder(URI.create(config.getUpdateUrl()))
         .POST(BodyPublishers.ofString("update=" + URLEncoder.encode(updateQuery,
                 StandardCharsets.UTF_8)))
-        .header(HttpHeaders.CONTENT_TYPE, Constants.APPLICATION_FORM_URLENCODED_VALUE)
+        .header(CONTENT_TYPE, Constants.APPLICATION_FORM_URLENCODED_VALUE)
         .build();
     System.out.println(request);
     execute(request, httpClient);
-  }
-
-  @Override
-  @Deprecated
-  public void upload(Model model, String graphUri) {
-    updateGraph(graphUri, model);
   }
 
   @Override
@@ -107,7 +120,7 @@ public class FusekiSparqlService implements SparqlService {
     final BodyPublisher p = BodyPublishers.ofByteArray(writer.toString().getBytes());
     final HttpRequest.Builder builder = HttpRequest
         .newBuilder(URI.create(insertUrl))
-        .header(HttpHeaders.CONTENT_TYPE, config.getTurtleMimeType() + ";charset=utf-8");
+        .header(CONTENT_TYPE, config.getTurtleMimeType() + ";charset=utf-8");
 
     final HttpRequest request = (replace ? builder.PUT(p) : builder.POST(p)).build();
     execute(request, httpClient);