+++ semanticz-connector-stardog/src/main/java/zone/cogni/semanticz/connectors/stardog/StardogSparqlService.java
@@ -1,6 +1,24 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 package zone.cogni.semanticz.connectors.stardog;
 
-import org.apache.http.HttpHeaders;
 import org.apache.jena.query.QueryExecution;
 import org.apache.jena.query.QueryExecutionBuilder;
 import org.apache.jena.query.ResultSet;
@@ -24,6 +42,8 @@ import java.net.http.HttpRequest.BodyPublisher;
 import java.nio.charset.StandardCharsets;
 import java.util.function.Function;
 
+import static zone.cogni.semanticz.connectors.utils.Constants.CONTENT_TYPE;
+
 public class StardogSparqlService implements SparqlService {
   private final String endpointUrl;
   private final HttpClient httpClient;
@@ -44,7 +64,7 @@ public class StardogSparqlService implements SparqlService {
       request = HttpRequest
           .newBuilder(URI.create(endpointUrl))
           .POST(HttpRequest.BodyPublishers.ofFile(file.toPath()))
-          .header(HttpHeaders.CONTENT_TYPE, Lang.TURTLE.getHeaderString()+";charset=utf-8")
+          .header(CONTENT_TYPE, Lang.TURTLE.getHeaderString()+";charset=utf-8")
           .build();
     } catch (FileNotFoundException e) {
       throw new RuntimeException(e);
@@ -66,7 +86,7 @@ public class StardogSparqlService implements SparqlService {
     final HttpRequest request = HttpRequest
         .newBuilder(URI.create(endpointUrl + "/update"))
         .POST(HttpRequest.BodyPublishers.ofString("update=" + URLEncoder.encode(updateQuery, StandardCharsets.UTF_8), StandardCharsets.UTF_8))
-        .header(HttpHeaders.CONTENT_TYPE, Constants.APPLICATION_FORM_URLENCODED_VALUE)
+        .header(CONTENT_TYPE, Constants.APPLICATION_FORM_URLENCODED_VALUE)
         .build();
     HttpClientUtils.execute(request, httpClient);
   }
@@ -78,11 +98,6 @@ public class StardogSparqlService implements SparqlService {
       // TODO check with empty default graph ((QueryEngineHTTP) queryExecution).setDefaultGraphURIs(Collections.emptyList());
       return queryExecution.execAsk();
     }
-  }
-
-  @Override
-  public void upload(Model model, String graphUri) {
-    updateGraph(graphUri, model);
   }
 
   @Override
@@ -99,7 +114,7 @@ public class StardogSparqlService implements SparqlService {
     String graphStoreUrl = endpointUrl + "?graph=" + URLEncoder.encode(graphUri, StandardCharsets.UTF_8);
     byte[] body = JenaUtils.toByteArray(model, TripleSerializationFormat.turtle);
     final HttpRequest.Builder builder = HttpRequest.newBuilder(URI.create(graphStoreUrl))
-        .header(HttpHeaders.CONTENT_TYPE, Lang.TURTLE.getHeaderString() + ";charset=utf-8");
+        .header(CONTENT_TYPE, Lang.TURTLE.getHeaderString() + ";charset=utf-8");
     final BodyPublisher p = HttpRequest.BodyPublishers.ofByteArray(body);
     final HttpRequest request = (replace ? builder.PUT(p) : builder.POST(p)).build();
     HttpClientUtils.execute(request, httpClient);