+++ semanticz-connector-virtuoso/src/main/java/zone/cogni/semanticz/connectors/virtuoso/VirtuosoApacheHttpClientRdfStoreService.java
@@ -1,3 +1,22 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 package zone.cogni.semanticz.connectors.virtuoso;
 
 import io.vavr.control.Try;
@@ -36,6 +55,9 @@ import java.io.StringWriter;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.concurrent.TimeUnit;
+
+import static zone.cogni.semanticz.connectors.utils.Constants.CONTENT_TYPE;
+import static zone.cogni.semanticz.connectors.utils.Constants.TEXT_TURTLE;
 
 /**
  * Implementation of VirtuosoRdfStoreService backed directly by Apache HttpClient.
@@ -101,7 +123,7 @@ public class VirtuosoApacheHttpClientRdfStoreService implements RdfStoreService 
     log.info("Calling {} with basic auth: {}", url, graphCrudUseBasicAuth);
     HttpEntityEnclosingRequestBase request = replace ? new HttpPut(url) : new HttpPost(url);
     request.setEntity(new ByteArrayEntity(data));
-    request.setHeader("Content-Type", "text/turtle;charset=utf-8");
+    request.setHeader(CONTENT_TYPE, TEXT_TURTLE + ";charset=utf-8");
     if (graphCrudUseBasicAuth) {
       request.setHeader("Authorization", "Basic " + Base64.encodeBase64String(
           (rdfStoreUser + ":" + rdfStorePassword).getBytes(StandardCharsets.UTF_8)));