+++ semanticz-connectors-common/src/main/java/zone/cogni/semanticz/connectors/utils/JenaUtils.java
@@ -1,11 +1,28 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 package zone.cogni.semanticz.connectors.utils;
 
-import org.apache.commons.io.FilenameUtils;
 import org.apache.jena.rdf.model.*;
 import org.apache.jena.riot.RDFLanguages;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import zone.cogni.semanticz.connectors.CognizoneException;
 
 import java.io.*;
 import java.util.*;
@@ -34,7 +51,7 @@ public class JenaUtils {
       outputStream.flush();
       return outputStream.toByteArray();
     } catch (IOException e) {
-      throw new CognizoneException(e);
+      throw new RuntimeException(e);
     }
   }
 
@@ -55,13 +72,6 @@ public class JenaUtils {
         log.warn("Closing model failed.", e);
       }
     }
-  }
-
-  public static String getLangByResourceName(String resourceName) {
-    String ext = FilenameUtils.getExtension(resourceName);
-    if (ext.equalsIgnoreCase("ttl")) return "TTL";
-    //TODO: add other types
-    return null;
   }
 
   public static Model readInto(File file, Model model) {