+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+1.1.0
+- added license headers
+- build scripts updated for publication to Nexus and Maven Central
+- fixing issues with `VirtuosoSparqlService` tests
+- added `VirtuosoApacheHttpClientRdfStoreServiceTest` tests
+- removing deprecated `CognizoneException` class and `SparqlService.queryForModel` and `SparqlService.upload` methods
+- adding `SparqlService.isEmptyGraph` method
+- parameterization of integration tests for Virtuoso and GraphDB
+
 1.0.0
 - extracted from ASquare